+++ src/main/java/kitchenpos/product/dto/response/ProductResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.dto.response;
+package kitchenpos.product.dto.response;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.doamin.Product;
 
 public class ProductResponse {
