+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.product.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.doamin.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
