+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -4,9 +4,9 @@ import java.net.URI;
 import java.util.List;
 import javax.validation.Valid;
 import kitchenpos.application.TableService;
+import kitchenpos.dto.request.TableCreateRequest;
 import kitchenpos.dto.request.TableUpdateEmptyRequest;
 import kitchenpos.dto.request.TableUpdateGuestRequest;
-import kitchenpos.dto.request.TableCreateRequest;
 import kitchenpos.dto.response.TableResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
