+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -9,11 +9,12 @@ import static org.mockito.Mockito.times;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.dao.ProductRepository;
+import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.domain.Price;
-import kitchenpos.domain.Product;
-import kitchenpos.dto.request.ProductCreateRequest;
-import kitchenpos.dto.response.ProductResponse;
+import kitchenpos.product.doamin.Product;
+import kitchenpos.product.dto.request.ProductCreateRequest;
+import kitchenpos.product.dto.response.ProductResponse;
+import kitchenpos.product.application.ProductService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
